# Conflicts:
#	src/pages/Space.tsx
#	src/pages/space/Settings.tsx
#	src/providers/SpaceProvider.tsx

+++ src/utils/number.ts
@@ -3,5 +3,5 @@ export const stringToNum = (input: string | undefined): number | undefined => {
     return undefined;
   }
 
-  return parseInt(input.replaceAll(',', ''));
+  return parseFloat(input.replaceAll(',', ''));
 };
